requirements_firewall.txt
redis
pyyaml
service
